pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "telegram-azkar-bot"
version = "2.0.0"
description = "Professional Islamic Azkar Telegram Bot with scheduled content delivery"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Omar Mahdy", email = "omarelmhdi@gmail.com"}
]
maintainers = [
    {name = "Omar Mahdy", email = "omarelmhdi@gmail.com"}
]
keywords = ["telegram", "bot", "islamic", "azkar", "prayer", "remembrance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Religion",
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "apscheduler>=3.11.0",
    "python-telegram-bot>=22.3",
    "pytz>=2025.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Mavdii/bot"
Repository = "https://github.com/Mavdii/bot"
Issues = "https://github.com/Mavdii/bot/issues"
Changelog = "https://github.com/Mavdii/bot/blob/main/CHANGELOG.md"

[project.scripts]
azkar-bot = "azkar_bot.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/azkar_bot"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/azkar_bot",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "telegram.*",
    "apscheduler.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/azkar_bot"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
